tests: wrap Python-int kwargs in jnp.int32 to satisfy strict pylcm types
Pylcm's typing aliases (Period, IntND) are now strict — Python ints no
longer satisfy them. Wrap kwarg literals in test_pensions,
test_pension_integration, test_social_security, test_ss_benefit_integration,
and test_model_components. Hoist top-level integer constants
(PERIOD, his/old_his/new_his) to jnp.int32. Convert _RATIO_NP to
pd.Series at the compute_di_dropout_scale call site.

aca-model: drop the leftover MAX_CONSUMPTION_UNEQUIV import + the
test_max_consumption_unequiv_attached_from_canonical_constant test
(the dynamic Model attribute is gone).

+++ tests/test_model_components.py
@@ -164,8 +164,8 @@ def test_next_aime_accrual() -> None:
     result = social_security.next_aime(
         aime=jnp.array(1000.0),
         labor_income=jnp.array(50000.0),
-        period=55,
-        age=55,
+        period=jnp.int32(55),
+        age=jnp.int32(55),
         benefit_withheld_fraction=jnp.array(0.0),
         earnings_test_credited_back=jnp.zeros(100),
         earnings_test_repealed_age=70,
